+++ Carbon.IIS/Functions/Get-CIisHttpRedirect.ps1
@@ -1,4 +1,11 @@
 
+enum CHttpResponseStatus
+{
+    Permanent = 301;
+    Found = 302;
+    Temporary = 307;
+}
+
 function Get-CIisHttpRedirect
 {
     <#
@@ -6,21 +13,20 @@ function Get-CIisHttpRedirect
     Gets the HTTP redirect settings for a website or virtual directory/application under a website.
     
     .DESCRIPTION
-    Returns a `Carbon.Iis.HttpRedirectConfigurationSection` object for the given HTTP redirect settings.  The object contains the following properties:
+    Returns a `[Microsoft.Web.Administration.ConfigurationSection]` object with these attributes:
     
-     * Enabled - `True` if the redirect is enabled, `False` otherwise.
-     * Destination - The URL where requests are directed to.
-     * HttpResponseCode - The HTTP status code sent to the browser for the redirect.
-     * ExactDestination - `True` if redirects are to destination, regardless of the request path.  This will send all requests to `Destination`.
-     * ChildOnly - `True` if redirects are only to content in the destination directory (not subdirectories).
-	 
-    Beginning with Carbon 2.0.1, this function is available only if IIS is installed.
+     * enabled - `True` if the redirect is enabled, `False` otherwise.
+     * destination - The URL where requests are directed to.
+     * httpResponseCode - The HTTP status code sent to the browser for the redirect.
+     * exactDestination - `True` if redirects are to destination, regardless of the request path.  This will send all
+     requests to `Destination`.
+     * childOnly - `True` if redirects are only to content in the destination directory (not subdirectories).
+
+     Use the `GetAttributeValue` and `SetAttributeValue` to get and set values and the `CommitChanges` method to persist
+     the changes to IIS.
 
     .LINK
     http://www.iis.net/configreference/system.webserver/httpredirect
-     
-    .OUTPUTS
-    Carbon.Iis.HttpRedirectConfigurationSection.
      
     .EXAMPLE
     Get-CIisHttpRedirect -SiteName ExampleWebsite 
@@ -33,25 +39,21 @@ function Get-CIisHttpRedirect
     Gets the redirect settings for the MyVirtualDirectory virtual directory under ExampleWebsite.
     #>
     [CmdletBinding()]
+    [OutputType([Microsoft.Web.Administration.ConfigurationSection])]
     param(
-        [Parameter(Mandatory=$true)]
-        [string]
         # The site's whose HTTP redirect settings will be retrieved.
-        $SiteName,
+        [Parameter(Mandatory)]
+        [String] $SiteName,
         
-        [Alias('Path')]
-        [string]
         # The optional path to a sub-directory under `SiteName` whose settings to return.
-        $VirtualPath = ''
+        [Alias('Path')]
+        [String] $VirtualPath = ''
     )
     
     Set-StrictMode -Version 'Latest'
-
     Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
     Get-CIisConfigurationSection -SiteName $SiteName `
-                                -VirtualPath $VirtualPath `
-                                -SectionPath 'system.webServer/httpRedirect' `
-                                -Type ([Carbon.Iis.HttpRedirectConfigurationSection])
+                                 -VirtualPath $VirtualPath `
+                                 -SectionPath 'system.webServer/httpRedirect'
 }
-