+++ Carbon.IIS/Functions/Get-CIisMimeMap.ps1
@@ -6,16 +6,15 @@ function Get-CIisMimeMap
     Gets the file extension to MIME type mappings.
     
     .DESCRIPTION
-    IIS won't serve static content unless there is an entry for it in the web server or website's MIME map configuration. This function will return all the MIME maps for the current server.  The objects returned are instances of the `Carbon.Iis.MimeMap` class, and contain the following properties:
+    IIS won't serve static content unless there is an entry for it in the web server or website's MIME map
+    configuration. This function will return all the MIME maps for the current server.  The objects returned have these
+    properties:
     
      * `FileExtension`: the mapping's file extension
      * `MimeType`: the mapping's MIME type
     
     Beginning with Carbon 2.0.1, this function is available only if IIS is installed.
 
-    .OUTPUTS
-    Carbon.Iis.MimeMap.
-    
     .LINK
     Set-CIisMimeMap
     
@@ -45,26 +44,21 @@ function Get-CIisMimeMap
     Gets all the file extension to MIME type mappings for the `DeathStar`'s `ExhausePort` directory.
     #>
     [CmdletBinding(DefaultParameterSetName='ForWebServer')]
-    [OutputType([Carbon.Iis.MimeMap])]
     param(
-        [Parameter(Mandatory=$true,ParameterSetName='ForWebsite')]
-        [string]
         # The website whose MIME mappings to return.  If not given, returns the web server's MIME map.
-        $SiteName,
+        [Parameter(Mandatory, ParameterSetName='ForWebsite')]
+        [String] $SiteName,
         
+        # The directory under the website whose MIME mappings to return.  Optional.
         [Parameter(ParameterSetName='ForWebsite')]
         [Alias('Path')]
-        [string]
-        # The directory under the website whose MIME mappings to return.  Optional.
-        $VirtualPath = '',
+        [String] $VirtualPath = '',
         
-        [string]
         # The name of the file extensions to return. Wildcards accepted.
-        $FileExtension = '*',
+        [String] $FileExtension = '*',
         
-        [string]
         # The name of the MIME type(s) to return.  Wildcards accepted.
-        $MimeType = '*'
+        [String] $MimeType = '*'
     )
 
     Set-StrictMode -Version 'Latest'
@@ -78,11 +72,14 @@ function Get-CIisMimeMap
         $getIisConfigSectionParams['VirtualPath'] = $VirtualPath
     }
 
-    $staticContent = Get-CIisConfigurationSection -SectionPath 'system.webServer/staticContent' @getIisConfigSectionParams
+    $staticContent =
+        Get-CIisConfigurationSection -SectionPath 'system.webServer/staticContent' @getIisConfigSectionParams
     $staticContent.GetCollection() | 
         Where-Object { $_['fileExtension'] -like $FileExtension -and $_['mimeType'] -like $MimeType } |
         ForEach-Object {
-            New-Object 'Carbon.Iis.MimeMap' ($_['fileExtension'],$_['mimeType'])
+            $mimeMap = [pscustomobject]@{ FileExtension = $_['fileExtension']; MimeType = $_['mimeType'] }
+            $mimeMap.pstypenames.Add('Carbon.Iis.MimeMap')
+            $mimeMap | Write-Output
         }
 }